Refactor: pass todos state to Header, Counter, and TodoList components
-"Lifting state up":
Moved the todos state from the TodoList (as it did not have any child components) to the main App component, and pass down as props to components that need it.

As other components needed the state which was not able to be accessed from the TodoList component (as it had no children).